+++ tests/test_mlpf_models.py
@@ -117,6 +117,53 @@ def test_mlpf_gnn_lsh():
     assert preds_pu.shape == (batch_size, seq_len, 2)
 
 
+def test_mlpf_litept():
+    try:
+        from mlpf.model.litept import LitePTLayer  # noqa: F401
+    except ImportError:
+        pytest.skip("LitePT not available")
+
+    if not torch.cuda.is_available():
+        pytest.skip("CUDA not available, LitePT (spconv) requires it")
+
+    config_dict = {
+        "dataset": "cms",
+        "data_dir": "/tmp",
+        "model": {
+            "type": "litept",
+            "litept": {
+                "num_convs": 1,
+                "embedding_dim": 16,
+                "width": 16,
+                "enc_channels": [16, 16, 16, 16, 16],
+                "dec_channels": [16, 16, 16, 16],
+            },
+        },
+        "conv_type": "litept",
+    }
+    config = MLPFConfig.model_validate(config_dict)
+    model = MLPF(config)
+
+    batch_size = 2
+    seq_len = 16
+    input_dim = config.input_dim
+
+    X = torch.randn(batch_size, seq_len, input_dim)
+    elem_types = torch.tensor(config.elemtypes_nonzero)
+    X[..., 0] = elem_types[torch.randint(0, len(elem_types), (batch_size, seq_len))].to(X.dtype)
+    X[..., 1] = torch.exp(X[..., 1])
+    X[..., 5] = torch.exp(X[..., 5])
+
+    mask = torch.ones(batch_size, seq_len, dtype=torch.bool)
+
+    preds_binary_particle, preds_pid, preds_momentum, preds_pu = model(X, mask)
+
+    assert preds_binary_particle.shape == (batch_size, seq_len, 2)
+    assert preds_pid.shape == (batch_size, seq_len, config.num_classes)
+    assert preds_momentum.shape == (batch_size, seq_len, 5)
+    assert preds_pu.shape == (batch_size, seq_len, 2)
+
+
 @pytest.mark.parametrize("dataset", ["cms", "clic", "cld"])
 def test_mlpf_datasets(dataset):
     config_dict = {
